RFC: Support list type messages which have a single object
Summary:
We had a user whose kafka messages were a single JSON object wrapped in an array (array of one element).
Updating our JsonParser to try deserializing the record as map first, it it fails then try deserializing as a list.
For now the parser fails hard if the list has 0 or more than 1 object.

Test Plan:
unit-tests and tested manually using the new jar in master. With the old jar I saw connect throwing the same excpetion
as what the user was seeing on their end.

Reviewers: #platform, kw, joe, dhruba

Reviewed By: #platform, kw, joe

Differential Revision: https://rockset.phacility.com/D9002